Remove meetings and update delegation of web duties
Update the Policy and Operations manual with the following changes.

Remove meetings and reflow how the proposal process works.

Update President and Secretary duties to reflect meetings not happening.

Remove the delegation of website duties to Logan per CCT agreement.

Update the Executive Action document to reflect the use of said authority when the Policy and Operations manual can't support a necessary change.